# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Founders Finance API
servers:
  - url: /api
    description: Base API path
...
security:
  - sessionCookie: []
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List accounts
      parameters:
        - name: entity_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a company account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Edit, deactivate, or reactivate an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      parameters:
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Edit, deactivate, or reactivate a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
  /expenses/{transactionId}/allocations:
    post:
      operationId: createExpenseAllocations
      tags: [expenses]
      summary: Set allocations for an expense
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationsBody"
      responses:
        "200":
          description: Allocations created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseAllocation"
  /allocation-presets:
    get:
      operationId: listAllocationPresets
      tags: [allocation-presets]
      summary: List allocation presets
      parameters:
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of allocation presets with lines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllocationPreset"
    post:
      operationId: createAllocationPreset
      tags: [allocation-presets]
      summary: Create an allocation preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationPresetBody"
      responses:
        "201":
          description: Created allocation preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationPreset"
  /allocation-presets/{id}:
    get:
      operationId: getAllocationPreset
      tags: [allocation-presets]
      summary: Get allocation preset by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Allocation preset with lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationPreset"
    put:
      operationId: updateAllocationPreset
      tags: [allocation-presets]
      summary: Edit, deactivate, or reactivate an allocation preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationPresetBody"
      responses:
        "200":
          description: Updated allocation preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationPreset"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ff_session
...
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        account_type:
          type: string
        institution_name:
          type: string
          nullable: true
        last_four:
          type: string
          nullable: true
        opening_balance:
          type: number
        current_balance:
          type: number
        is_tax_reserve:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - account_type
        - opening_balance
        - current_balance
        - is_tax_reserve
        - is_active
        - created_at
        - updated_at
    CreateAccountBody:
      type: object
      properties:
        entity_id:
          type: string
          format: uuid
        name:
          type: string
        account_type:
          type: string
          enum: [checking, savings, credit_card, cash, loan, other]
        institution_name:
          type: string
          nullable: true
        last_four:
          type: string
          pattern: "^[0-9]{4}$"
          nullable: true
        opening_balance:
          type: number
        is_tax_reserve:
          type: boolean
      required: [entity_id, name, account_type]
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        account_type:
          type: string
          enum: [checking, savings, credit_card, cash, loan, other]
        institution_name:
          type: string
          nullable: true
        last_four:
          type: string
          pattern: "^[0-9]{4}$"
          nullable: true
        is_tax_reserve:
          type: boolean
        is_active:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category_type:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
      required:
        - id
        - name
        - category_type
        - is_active
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        category_type:
          type: string
          enum: [expense, income, asset, liability, equity, other]
        description:
          type: string
          nullable: true
      required: [name, category_type]
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        category_type:
          type: string
          enum: [expense, income, asset, liability, equity, other]
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
...
    ExpenseAllocation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transaction_id:
          type: string
          format: uuid
        target_entity_id:
          type: string
          format: uuid
        entity_short_code:
          type: string
          nullable: true
        entity_display_name:
          type: string
          nullable: true
        entity_primary_color:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        allocation_percent:
          type: number
          nullable: true
        allocation_amount:
          type: number
        memo:
          type: string
          nullable: true
        creates_intercompany_balance:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - transaction_id
        - target_entity_id
        - allocation_amount
        - creates_intercompany_balance
        - created_at
...
    AllocationInput:
      type: object
      properties:
        target_entity_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        allocation_percent:
          type: number
          nullable: true
        allocation_amount:
          type: number
        memo:
          type: string
          nullable: true
        creates_intercompany_balance:
          type: boolean
      required:
        - target_entity_id
        - allocation_amount
        - creates_intercompany_balance
...
    CreateAllocationsBody:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/AllocationInput"
      required:
        - allocations
    AllocationPresetLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        preset_id:
          type: string
          format: uuid
        entity_id:
          type: string
          format: uuid
        entity_short_code:
          type: string
          nullable: true
        entity_display_name:
          type: string
          nullable: true
        entity_primary_color:
          type: string
          nullable: true
        percent:
          type: number
      required:
        - id
        - preset_id
        - entity_id
        - percent
    AllocationPreset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        lines:
          type: array
          items:
            $ref: "#/components/schemas/AllocationPresetLine"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - is_active
        - lines
        - created_at
        - updated_at
    AllocationPresetLineInput:
      type: object
      properties:
        entity_id:
          type: string
          format: uuid
        percent:
          type: number
          exclusiveMinimum: 0
          maximum: 100
      required: [entity_id, percent]
    CreateAllocationPresetBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AllocationPresetLineInput"
      required: [name, lines]
    UpdateAllocationPresetBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AllocationPresetLineInput"
...